backend/requirements.txt
Flask==2.0.1
Flask-Cors==5.0.0
requests==2.25.1
Werkzeug==2.2.2  # Specify a compatible version
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
psycopg2-binary
gunicorn